RSA re-refactor
